+++ src/count_kmers_specific/count_kmers_numeric_matrix.h
@@ -1,5 +1,5 @@
-#ifndef SEQR_FIND_KMERS_INTEGER_MATRIX_H
-#define SEQR_FIND_KMERS_INTEGER_MATRIX_H
+#ifndef SEQR_COUNT_KMERS_NUMERIC_MATRIX_H
+#define SEQR_COUNT_KMERS_NUMERIC_MATRIX_H
 
 #include <Rcpp.h>
 #include <vector>
@@ -7,18 +7,20 @@
 #include "../dictionary/unordered_map_wrapper.h"
 #include "../kmer_counting_result.h"
 #include "../kmer_task_solver.h"
-#include "../common_config.h"
 
-inline InputToStringItemConverter_t<int> getIntToStringConverter() {
-    return [](const int &elem) -> std::string {
-        return std::to_string(elem);
+inline InputToStringItemConverter_t<double> getDoubleToStringConverter(int decimalPrecision) {
+    return [decimalPrecision](const double &elem) -> std::string {
+        std::ostringstream stream;
+        stream << std::fixed << std::setprecision(decimalPrecision);
+        stream << elem;
+        return stream.str();
     };
 }
 
 template<class algorithm_params_t>
 inline
-Rcpp::List findKMersSpecific(Rcpp::IntegerMatrix &sequenceMatrix,
-                             Rcpp::IntegerVector &alphabet,
+Rcpp::List countKMersSpecific(Rcpp::NumericMatrix &sequenceMatrix,
+                             Rcpp::NumericVector &alphabet,
                              std::vector<int> &gaps,
                              bool positionalKMers,
                              bool withKMerCounts,
@@ -27,23 +29,23 @@ Rcpp::List findKMersSpecific(Rcpp::IntegerMatrix &sequenceMatrix,
                              bool verbose,
                              bool parallelMode,
                              algorithm_params_t &algorithmParams) {
-    using encoded_elem_t = config::encoded_elem_t ;
+    using encoded_elem_t = config::encoded_elem_t;
     auto alphabetEncoding = std::move(
-            alphabetEncoding::getDefaultAlphabetEncoder<Rcpp::IntegerVector, int, encoded_elem_t, dictionary::UnorderedMapWrapper>(alphabet));
+            alphabetEncoding::getDefaultAlphabetEncoder<Rcpp::NumericVector, double, encoded_elem_t, dictionary::UnorderedMapWrapper>(alphabet));
 
     auto batchFunc = [&](KMerCountingResult &kMerCountingResult, int seqBegin, int seqEnd) {
-        KMerTaskConfig<RcppParallel::RMatrix<int>::Row, int> kMerTaskConfig(
+        KMerTaskConfig<RcppParallel::RMatrix<double>::Row, double> kMerTaskConfig(
                 (seqEnd - seqBegin),
-                getRMatrixRowGetter<Rcpp::IntegerMatrix, decltype(alphabetEncoding)::input_elem_t>(sequenceMatrix, seqBegin),
+                getRMatrixRowGetter<Rcpp::NumericMatrix, decltype(alphabetEncoding)::input_elem_t>(sequenceMatrix, seqBegin),
                 gaps,
                 positionalKMers,
                 withKMerCounts,
                 parallelMode,
-                getIntToStringConverter(),
+                getDoubleToStringConverter(3),
                 config::DEFAULT_KMER_ITEM_SEPARATOR,
                 config::DEFAULT_KMER_SECTION_SEPARATOR);
         updateKMerCountingResult<
-                RcppParallel::RMatrix<int>::Row,
+                RcppParallel::RMatrix<double>::Row,
                 decltype(alphabetEncoding)::input_elem_t,
                 decltype(alphabetEncoding),
                 algorithm_params_t>(kMerTaskConfig,
@@ -56,4 +58,4 @@ Rcpp::List findKMersSpecific(Rcpp::IntegerMatrix &sequenceMatrix,
     return computeKMersInBatches(batchFunc, sequenceMatrix.nrow(), batchSize, verbose);
 }
 
-#endif //SEQR_FIND_KMERS_INTEGER_MATRIX_H
+#endif //SEQR_COUNT_KMERS_NUMERIC_MATRIX_H
