+++ src/count_kmers_specific/rmatrix_row_getter.h
