+++ src/count_kmers_specific/safe_sequences_wrapper.h
